Add support to render relationships and docstrings in snapshots
