+++ snapshots/output/multi-project/packages/a/src/index.ts
@@ -1,5 +1,7 @@
   export function a(): string {
 //                ^ definition @example/a 1.0.0 src/`index.ts`/a().
+//                documentation ```ts\n() => string\n```
+//                documentation 
     return ''
   }
