+++ snapshots/output/multi-project/packages/b/src/b.ts
@@ -3,6 +3,8 @@
   
   export function b() {
 //                ^ definition @example/b 1.0.0 src/`b.ts`/b().
+//                documentation ```ts\n() => string\n```
+//                documentation 
     return a()
 //         ^ reference @example/a 1.0.0 src/`index.ts`/a().
   }
