+++ snapshots/output/pure-js/src/main.js
@@ -1,6 +1,10 @@
   function fib(n) {
 //         ^^^ definition pure-js 1.0.0 src/`main.js`/fib().
+//         documentation ```ts\n(n: any) => any\n```
+//         documentation 
 //             ^ definition pure-js 1.0.0 src/`main.js`/fib().(n)
+//             documentation ```ts\nany\n```
+//             documentation 
     if (n <= 1) {
 //      ^ reference pure-js 1.0.0 src/`main.js`/fib().(n)
       return 0
@@ -14,7 +18,11 @@
   
   function print_fib(a) {
 //         ^^^^^^^^^ definition pure-js 1.0.0 src/`main.js`/print_fib().
+//         documentation ```ts\n(a: any) => void\n```
+//         documentation 
 //                   ^ definition pure-js 1.0.0 src/`main.js`/print_fib().(a)
+//                   documentation ```ts\nany\n```
+//                   documentation 
     console.log(fib(a))
 //  ^^^^^^^ reference typescript 4.6.2 lib/`lib.dom.d.ts`/console.
 //  ^^^^^^^ reference @types/node 17.0.14 `globals.d.ts`/console.
@@ -28,30 +36,44 @@
   
   var y = 'Hello'
 //    ^ definition pure-js 1.0.0 src/`main.js`/y.
+//    documentation ```ts\nstring\n```
+//    documentation 
   function capture() {
 //         ^^^^^^^ definition pure-js 1.0.0 src/`main.js`/capture().
+//         documentation ```ts\n() => string\n```
+//         documentation 
     return y
 //         ^ reference pure-js 1.0.0 src/`main.js`/y.
   }
   const capture_lambda = () => {
 //      ^^^^^^^^^^^^^^ definition pure-js 1.0.0 src/`main.js`/capture_lambda.
+//      documentation ```ts\n() => string\n```
+//      documentation 
     return y
 //         ^ reference pure-js 1.0.0 src/`main.js`/y.
   }
   
   for (var i = 0; i <= 10; i++) {}
 //         ^ definition local 2
+//         documentation ```ts\nnumber\n```
+//         documentation 
 //                ^ reference local 2
 //                         ^ reference local 2
   
   for (const x of [1, 2, 3]) {
 //           ^ definition local 5
+//           documentation ```ts\nnumber\n```
+//           documentation 
   }
   
   var a = 0
 //    ^ definition pure-js 1.0.0 src/`main.js`/a.
+//    documentation ```ts\nnumber\n```
+//    documentation 
   var a = 1
 //    ^ definition pure-js 1.0.0 src/`main.js`/a.
+//    documentation ```ts\nnumber\n```
+//    documentation 
   print_fib(a)
 //^^^^^^^^^ reference pure-js 1.0.0 src/`main.js`/print_fib().
 //          ^ reference pure-js 1.0.0 src/`main.js`/a.
@@ -59,22 +81,30 @@
   
   function forever() {
 //         ^^^^^^^ definition pure-js 1.0.0 src/`main.js`/forever().
+//         documentation ```ts\n() => any\n```
+//         documentation 
     return forever()
 //         ^^^^^^^ reference pure-js 1.0.0 src/`main.js`/forever().
   }
   
   function use_before_def() {
 //         ^^^^^^^^^^^^^^ definition pure-js 1.0.0 src/`main.js`/use_before_def().
+//         documentation ```ts\n() => void\n```
+//         documentation 
     print_fib(n)
 //  ^^^^^^^^^ reference pure-js 1.0.0 src/`main.js`/print_fib().
 //            ^ reference local 8
     var n = 10
 //      ^ definition local 8
+//      documentation ```ts\nnumber\n```
+//      documentation 
   
     if (forever()) {
 //      ^^^^^^^ reference pure-js 1.0.0 src/`main.js`/forever().
       var m = 10
 //        ^ definition local 11
+//        documentation ```ts\nnumber\n```
+//        documentation 
     }
     print_fib(m)
 //  ^^^^^^^^^ reference pure-js 1.0.0 src/`main.js`/print_fib().
@@ -83,12 +113,18 @@
   
   function var_function_scope() {
 //         ^^^^^^^^^^^^^^^^^^ definition pure-js 1.0.0 src/`main.js`/var_function_scope().
+//         documentation ```ts\n() => void\n```
+//         documentation 
     var k = 0
 //      ^ definition local 14
+//      documentation ```ts\nnumber\n```
+//      documentation 
     if (forever()) {
 //      ^^^^^^^ reference pure-js 1.0.0 src/`main.js`/forever().
       var k = 1
 //        ^ definition local 14
+//        documentation ```ts\nnumber\n```
+//        documentation 
     }
     print_fib(k)
 //  ^^^^^^^^^ reference pure-js 1.0.0 src/`main.js`/print_fib().
