+++ snapshots/output/react/src/MyTSXElement.tsx
@@ -3,9 +3,13 @@
   
   export interface MyProps {}
 //                 ^^^^^^^ definition react-example 1.0.0 src/`MyTSXElement.tsx`/MyProps#
+//                 documentation ```ts\nMyProps\n```
+//                 documentation 
   
   export const MyTSXElement: React.FunctionComponent<MyProps> = ({}) => (<p></p>)
 //             ^^^^^^^^^^^^ definition react-example 1.0.0 src/`MyTSXElement.tsx`/MyTSXElement.
+//             documentation ```ts\nFunctionComponent<MyProps>\n```
+//             documentation 
 //                           ^^^^^ reference @types/react 17.0.0 `index.d.ts`/React/
 //                                 ^^^^^^^^^^^^^^^^^ reference @types/react 17.0.0 `index.d.ts`/React/FunctionComponent#
 //                                                   ^^^^^^^ reference react-example 1.0.0 src/`MyTSXElement.tsx`/MyProps#
