+++ snapshots/output/syntax/src/function.ts
@@ -1,3 +1,5 @@
   export function newFunction(): void {}
 //                ^^^^^^^^^^^ definition syntax 1.0.0 src/`function.ts`/newFunction().
+//                documentation ```ts\n() => void\n```
+//                documentation
