+++ snapshots/output/syntax/src/import.ts
@@ -10,6 +10,8 @@
   
   export function useEverything(): string {
 //                ^^^^^^^^^^^^^ definition syntax 1.0.0 src/`import.ts`/useEverything().
+//                documentation ```ts\n() => string\n```
+//                documentation 
     return (
       new Class('a').classProperty +
 //        ^^^^^ reference syntax 1.0.0 src/`class.ts`/Class#
