+++ snapshots/output/syntax/src/module.d.ts
@@ -1,5 +1,7 @@
   declare module 'a:b' {
     function hello(): string
 //           ^^^^^ definition syntax 1.0.0 src/`module.d.ts`/`'a:b'`/hello().
+//           documentation ```ts\n() => string\n```
+//           documentation 
   }
