+++ src/main.test.ts
@@ -9,9 +9,7 @@ import { test } from 'uvu'
 import { Input } from './Input'
 import * as lsif from './lsif'
 import { indexCommand } from './main'
-import { Range } from './Range'
-
-const lsiftyped = lsif.lib.codeintel.lsiftyped
+import { formatSnapshot } from './SnapshotTesting'
 
 function isUpdateSnapshot(): boolean {
   return process.argv.includes('--update-snapshots')
@@ -101,57 +99,6 @@ for (const snapshotDirectory of snapshotDirectories) {
       }
     }
   })
-}
-
-function formatSnapshot(
-  input: Input,
-  document: lsif.lib.codeintel.lsiftyped.Document
-): string {
-  const out: string[] = []
-  document.occurrences.sort(occurrencesByLine)
-  let occurrenceIndex = 0
-  for (const [lineNumber, line] of input.lines.entries()) {
-    out.push('  ')
-    out.push(line)
-    out.push('\n')
-    while (
-      occurrenceIndex < document.occurrences.length &&
-      document.occurrences[occurrenceIndex].range[0] === lineNumber
-    ) {
-      const occurrence = document.occurrences[occurrenceIndex]
-      occurrenceIndex++
-      if (occurrence.range.length > 3) {
-        // Skip multiline occurrences for now.
-        continue
-      }
-      const range = Range.fromLsif(occurrence.range)
-      out.push('//')
-      out.push(' '.repeat(range.start.character))
-      const length = range.end.character - range.start.character
-      if (length < 0) {
-        throw new Error(input.format(range, 'negative length occurrence!'))
-      }
-      out.push('^'.repeat(length))
-      out.push(' ')
-      const isDefinition =
-        (occurrence.symbol_roles & lsiftyped.SymbolRole.Definition) > 0
-      out.push(isDefinition ? 'definition' : 'reference')
-      out.push(' ')
-      const symbol = occurrence.symbol.startsWith('lsif-typescript npm ')
-        ? occurrence.symbol.slice('lsif-typescript npm '.length)
-        : occurrence.symbol
-      out.push(symbol)
-      out.push('\n')
-    }
-  }
-  return out.join('')
-}
-
-function occurrencesByLine(
-  a: lsif.lib.codeintel.lsiftyped.Occurrence,
-  b: lsif.lib.codeintel.lsiftyped.Occurrence
-): number {
-  return Range.fromLsif(a.range).compare(Range.fromLsif(b.range))
 }
 
 test.run()
